pyproject.toml
[tool.poetry]
name = "repl-nix-bytebin-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
werkzeug = "^3.0.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
